feat: CreateListを完成。これでargvからスタックに数を入れられるようになった。また、スタックの操作コマンドとしてAddNodeとDeleteNodeを作成

+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tsuetsug <tsuetsug@student.42tokyo.>       +#+  +:+       +#+        */
+/*   By: tsuetsug <tsuetsug@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/30 22:18:03 by tsuetsug          #+#    #+#             */
-/*   Updated: 2020/11/30 23:12:14 by tsuetsug         ###   ########.fr       */
+/*   Updated: 2021/08/10 18:10:34 by tsuetsug         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ typedef struct	s_split
 	size_t length;
 }				t_split;
 
-int				ft_atoi(const char *str);
+long			ft_atoi(const char *str);
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
 int				ft_isascii(int c);
